\name{create.xpose.plot.classes}
\alias{create.xpose.plot.classes}
\alias{print,xpose.multiple.plot-method}
\title{
Create xpose.multiple.plot class.
}
\description{
Creates a class for viewing and plotting xpose plots with multiple plots
on the same page or multiple pages.
}
\usage{
create.xpose.plot.classes()
}
\details{
}
\value{
}
\references{
}
\author{
Niclas Jonsson and Andrew C. Hooker
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ methods }

