\name{print.xpose.multiple.plot}
\alias{print.xpose.multiple.plot}
\title{
Print an Xpose multiple plot object.
}
\description{
Print an Xpose multiple plot object, which is the output from the
function \code{\link{xpose.multiple.plot}}.
}
\usage{
print.xpose.multiple.plot(x, ...)
}
\arguments{
  \item{x}{
Output object from the function \code{\link{xpose.multiple.plot}}.
}
  \item{\dots}{
Additional options passed to function. 
}
}
\details{
Print method for a plot class.
}
\value{
}
\references{
}
\author{
Niclas Jonsson and Andrew C. Hooker
}
\note{
}

\seealso{
  \code{\link{xpose.multiple.plot}}.
}
\examples{
}
\keyword{ methods }
