\name{xpose.create.title.text}
\alias{xpose.create.title.text}
\title{
Create Xpose title text for plots.
}
\description{
Create Xpose title text for plots.
}
\usage{
xpose.create.title.text(x, y, text, object, subset, text2 = NULL, ...)
}
\arguments{
  \item{x}{
The x-axis vaiable name.
}
  \item{y}{
The y-axis vaiable name.
}
  \item{text}{
Initial text in title. 
}
  \item{object}{
Xpose data object \code{\link[xpose4data]{xpose.data}}.
}
  \item{subset}{
Subset definition.
}
  \item{text2}{
Text at the end of the title.
}
  \item{\dots}{
Additional options passed to function.
}
}
\details{
}
\value{
}
\references{
}
\author{
Andrew C. Hooker
}
\note{
}

\seealso{
}
\examples{}
\keyword{ methods }
