\name{xpose.string.print}
\alias{xpose.string.print}
\title{
Print a pretty string.
}
\description{
Print a string with a certain number of characters per row.
}
\usage{
xpose.string.print(value, fill = 60, file = "")
}
\arguments{
  \item{value}{
The text to print.
}
  \item{fill}{
How wide should the text be per row.
}
  \item{file}{
Where to print. \code{""} means to the screen. 
}
}
\details{
}
\value{
}
\references{
}
\author{
Niclas Jonsson and Andrew C. Hooker
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ methods }
